+++ src/lib.rs
@@ -205,14 +205,13 @@ macro_rules! try_const {
 pub(crate) use try_const;
 
 use raw_bmp::ColorType;
-use raw_iter::{RawColors, Rle4Pixels, Rle8Pixels};
 
 pub use color_table::ColorTable;
 pub use header::CompressionMethod;
 pub use header::{Bpp, ChannelMasks, Header, RowOrder};
 pub use iter::Pixels;
 pub use raw_bmp::RawBmp;
-pub use raw_iter::{RawPixel, RawPixels};
+pub use raw_iter::{DynamicRawColors, RawColors, RawPixel, RawPixels, Rle4Colors, Rle8Colors};
 
 /// A BMP-format bitmap.
 ///
@@ -294,19 +293,20 @@ where
                 let fallback_color = C::from(Rgb888::BLACK);
                 if let Some(color_table) = self.raw_bmp.color_table() {
                     if header.compression_method == CompressionMethod::Rle4 {
-                        let mut colors = Rle4Pixels::new(&self.raw_bmp).map(|raw_pixel| {
+                        let mut colors = Rle4Colors::new(&self.raw_bmp);
+                        let map_color = |color: RawU4| {
                             color_table
-                                .get(raw_pixel.color)
+                                .get(color.into_inner() as u32)
                                 .map(Into::into)
                                 .unwrap_or(fallback_color)
-                        });
+                        };
                         // RLE produces pixels in bottom-up order, so we draw them line by line rather than the entire bitmap at once.
                         for y in (0..area.size.height).rev() {
+                            colors.start_row();
+
                             let row = Rectangle::new(Point::new(0, y as i32), slice_size);
-                            target.fill_contiguous(
-                                &row,
-                                colors.by_ref().take(area.size.width as usize),
-                            )?;
+                            let colors = colors.by_ref().map(map_color);
+                            target.fill_contiguous(&row, colors.take(area.size.width as usize))?;
                         }
                         Ok(())
                     } else {
@@ -328,19 +328,20 @@ where
                 let fallback_color = C::from(Rgb888::BLACK);
                 if let Some(color_table) = self.raw_bmp.color_table() {
                     if header.compression_method == CompressionMethod::Rle8 {
-                        let mut colors = Rle8Pixels::new(&self.raw_bmp).map(|raw_pixel| {
+                        let mut colors = Rle8Colors::new(&self.raw_bmp);
+                        let map_color = |color: RawU8| {
                             color_table
-                                .get(raw_pixel.color)
+                                .get(color.into_inner() as u32)
                                 .map(Into::into)
                                 .unwrap_or(fallback_color)
-                        });
+                        };
                         // RLE produces pixels in bottom-up order, so we draw them line by line rather than the entire bitmap at once.
                         for y in (0..area.size.height).rev() {
+                            colors.start_row();
+
                             let row = Rectangle::new(Point::new(0, y as i32), slice_size);
-                            target.fill_contiguous(
-                                &row,
-                                colors.by_ref().take(area.size.width as usize),
-                            )?;
+                            let colors = colors.by_ref().map(map_color);
+                            target.fill_contiguous(&row, colors.take(area.size.width as usize))?;
                         }
                         Ok(())
                     } else {